% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_manual.R
\name{make_manual}
\alias{make_manual}
\title{Produce the pdf manual for a package}
\usage{
make_manual(package_location = getwd(), output_file = NULL)
}
\arguments{
\item{package_location}{root directory for the package being documented}

\item{output_file}{file path for output}
}
\value{
The return value is an error code (0 for success), given the
invisible attribute (so needs to be printed explicitly). If the command
could not be run for any reason, the value is 127 and a warning is issued
(as from R 3.5.0). Otherwise the value is the exit status returned by the
command.
}
\description{
Produce the pdf manual for a package
}
