% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_local.R
\name{publish_local}
\alias{publish_local}
\title{Create or update a local repository}
\usage{
publish_local(repository_path, package_path = ".")
}
\arguments{
\item{repository_path}{Folder in which to create the repository or update an existing one.}

\item{package_path}{Path to the parent directory of the package to be published.}
}
\description{
Generate or update a directory structure that can be read by \code{utils::install.packages}.
}
\details{
This function creates a system of subdirectories
}
\examples{
\dontrun{publish_local()}
}
