% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_local.R
\name{publish_local}
\alias{publish_local}
\title{Create or update a local repository}
\usage{
publish_local(repository_path, package_path = ".")
}
\arguments{
\item{repository_path}{Folder in which to create the repository or update an existing one.}

\item{package_path}{Path to the parent directory of the package to be published.}
}
\description{
Generate or update a directory structure that can be read by \code{utils::install.packages}.
}
\details{
Create or update a system of subdirectories that is recognized by \code{utils::install.packages} as a
package repository. The function publishes the referenced package both as a source tarball and as
a binary package under the current operating system. As a result, installation of the package
will fail if attempted from a different operating system with no available toolchain unless
a separate binary is created and added.
}
\examples{
\dontrun{publish_local("C:/Path/To/Repo")}
}
